Separate functions into classes thematically
